+++ tests/python_side/parser/constant_type.py
@@ -1,7 +1,7 @@
 from unittest import TestCase, main
 
 from autocxxpy.core import CXXParser
-from autocxxpy.core.types.cxx_types import is_const_type
+from autocxxpy.core.core_types.cxx_types import is_const_type
 
 
 class ConstantType(TestCase):